shadow/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.3
pyyaml==6.0.1
python-dotenv==1.0.1
httpx==0.26.0
numpy==1.24.3

# LLM provider dependencies (optional)
openai==1.12.0
anthropic==0.18.0

# Vector operations for memory system
scikit-learn==1.3.0
